# Optimization for parsing single-character version components