--- a/src/libraries/System.Private.CoreLib/src/System/Version.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Version.cs
@@ -354,6 +354,15 @@ namespace System
 
         private static bool TryParseComponent(ReadOnlySpan<char> component, string componentName, bool throwOnFailure, out int parsedComponent)
         {
+            if(component.Length == 1)
+            {
+                var value = component[0] - '0';
+                if(value >= 0 && value <= 9)
+                {
+                    parsedComponent = value;
+                    return true;
+                }
+            }
             if (throwOnFailure)
             {
                 if ((parsedComponent = int.Parse(component, NumberStyles.Integer, CultureInfo.InvariantCulture)) < 0)