@jkotas 2021-04-12
I do not think that this optimization is worth the extra code.