@jkotas 2021-04-12
> It is very common for Version components to be a single digit number.

It is very common for integer parsing to be used for single digit numbers. It does not make sense everybody to manually special case it to get good performance. We should be working on improving the integer parsing instead.