@MarkPflug 2021-04-12
@jkotas I agree, and I started a discussion about this last Friday #51010, weighing the cost/benefit. Version parsing seemed like it would obviously benefit for very little effort, and had no downside. I didn't realize the few extra lines of code would be considered a problem.