@jkotas 2021-04-12
> I didn't realize the few extra lines of code would be considered a problem.

It setting bad precedent. There are many places that call the integer parsing methods in this repo and it is likely that many of those can be measured just like Version parsing. If we accept this change for version parsing, do we accept the same change for all of them?